Add DeoptimizeObjectsALotThread support
